requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
strawberry-graphql-django==0.2.5
